rename: myFetch 관련 함수 폴더 이동
